membuat new class yang tidak masuk di class employee
