+++ src/lib/Boss.java
@@ -1,0 +1,38 @@
+package lib
+
+import java.time.LocalDate;
+import java.time.Month;
+import java.util.LinkedList;
+import java.util.List;
+public class Boss{
+    /**
+	 * Fungsi untuk menentukan gaji bulanan pegawai berdasarkan grade kepegawaiannya (grade 1: 3.000.000 per bulan, grade 2: 5.000.000 per bulan, grade 3: 7.000.000 per bulan)
+	 * Jika pegawai adalah warga negara asing gaji bulanan diperbesar sebanyak 50%
+	 */
+    public void setMonthlySalary(int grade) {	
+		if (grade == 1) {
+			monthlySalary = 3000000;
+			if (isForeigner) {
+				monthlySalary = (int) (3000000 * 1.5);
+			}
+		}else if (grade == 2) {
+			monthlySalary = 5000000;
+			if (isForeigner) {
+				monthlySalary = (int) (3000000 * 1.5);
+			}
+		}else if (grade == 3) {
+			monthlySalary = 7000000;
+			if (isForeigner) {
+				monthlySalary = (int) (3000000 * 1.5);
+			}
+		}
+	}
+    public void setAnnualDeductible(int deductible) {	
+		this.annualDeductible = deductible;
+	}
+
+	public void setAdditionalIncome(int income) {	
+		this.otherMonthlyIncome = income;
+	}76
+
+}
